refactor(library/init): alternative definition for mem
The new definition prevents the generation of higher order unification
problems. However, it forces us to provide more information.
For example, the following definition cannot be elaborated

     lemma not_mem_nil (a : α) : a ∉ [] := ...

The new `mem` does not restrict [] to have type (list α).
One alternative is to fire type class resolution even when the type
contains metavariables (like Coq). Then, we would try to find an
instance for

             (has_mem α (list ?m))

and since we have the instance

             (has_mem α (list α))

it would assign ?m to α.

@jroesch Do you know how Rust handles this kind of problem?
When does Rust trigger type class resolution?

@avigad, @gebner and @kha: any ideas?