refactor(library/init/data): add fn_array
`fn_array A n` is defined as `fin n -> A`

In Lean, we currently don't have good support for using nested inductive
datatypes such as

   inductive Expr
   | app : string -> list Expr -> Expr
   | var : string -> Expr

The main issue is that users currently need to use well-founded
recursion to be able to define recursive functions over nested inductive
datatypes. Since the equation compiler does not have support for
well-founded recursion, this is a very laborious step.

This feature is needed in the KreMLin project.
So, I suggested they use the following workaround.

   inductive Expr
   | app : string -> forall n, fn_array Expr n -> Expr
   | var : string -> Expr

This is similar to the trick we used to use in the tactic framework.
Note that, `array Expr n` cannot be used because it would trigger the
nested inductive compiler.

After adding fn_array, I defined array on top of it.

Unfortunately, this change affected the hash_map module.
I repaired some of the broken lemmas, but we still have two broken ones.